MatchContactInfo Finished(Constraints Need Adjusting)
The overall layout of the screen is finished. I couldn't get the layout to match the drawing exactly so I chose to adapt it so that it's more consistent with the look of the profile view. The only thing I couldn't get right were the constraints for the pictures at the center. I also wasn't sure how to implement the tab bar at the bottom so that it matches the other screens.